+++ src/call/callHandler.ts
@@ -1,20 +1,24 @@
 import { CategoryChannel, Guild, TextChannel, User, VoiceChannel, VoiceState } from "discord.js";
+import { client } from "../../bot";
 
 const Discord = require('discord.js')
-const bot = new Discord.Client();
 
-let tempChannels : VoiceChannel[];
-bot.on('voiceStateUpdate', (fromState: VoiceState, state: VoiceState) => {
-    console.log(state.channel.toString());
-    if (state.channel.toString() == "Create Call") {
+
+let tempChannels: string[] = [];
+// move to call handler later
+let handleJoin = (fromState: VoiceState, state: VoiceState) => {
+    if (state.channel != null && state.channel.name == "Create Call") {
         createChannel(state);
     }
-    if (tempChannels.includes(fromState.channel)) {
-        if (fromState.channel.members.size == 0) {
+    
+    if (fromState.channel != undefined) {
+        if (tempChannels.includes(fromState.channelID) && fromState.channel.members.size == 0) {
+            tempChannels.splice(tempChannels.indexOf(fromState.channelID), 1);
             fromState.channel.delete();
         }
     }
-})
+    
+}
 
 function createChannel(state: VoiceState) {
     let category: CategoryChannel = state.channel.parent;
@@ -23,9 +27,12 @@ function createChannel(state: VoiceState) {
     
     guild.channels.create(`${user.username.toString()}'s Private Channel`, {type: 'voice'})
         .then(newChannel => {
-            tempChannels.push(newChannel);
+            tempChannels.push(newChannel.id);
+            newChannel.setParent(category);
+            state.member.voice.setChannel(newChannel);
         })
         .catch(console.log);
+    
 }
 
-
+module.exports = handleJoin;
